Basic water added
It's not perfect, but a simple "raise water" and "lower water" command
has been added, and mapped to the "J" and "K" keys.

With just over an hour left, I think this is as far as I can go in terms
of the program itself (although there are of course extremely messy
components still in the code).

Next goal is to set up a small Dockerfile if possible, and construct a
README.md